Fix and update legacy subcommand
The `--migrate` option has been removed since this is the only action
the command performs. Migrated fields are no published if they were
prior.

The `--exclude-orphans` flag has been removed and are simply not
migrated. The data model must exist in order to perform the heuristics,
so they should be migrated when the data model is available.