Correctly support SELECT DISTINCT queries without ORDER BY columns
Most database implementations of SELECT DISTINCT uses GROUP BY
under the hood per aggregate based on the columns in SELECT. As
a result, using columns in ORDER BY that do not appear in the
SELECT clause causes an error.

The solution for this is to silently include the the fields in
the SELECT statement if the query is using a DISTINCT.

The assumption here is that if a raw query is being generated,
this would be the normal (and required) behavior of a hand-written
SQL query. However, if used with an exporter, the extra columns will
be trimmed from the output due to the `force_distinct` parameter in
`Exporter.read()` which is `True` by default. This provides the
benefit of database ordering, but with the extra output.