Remove unnecessary conditionals.
These if statements are redundant because we have an assert checking
that the list lengths match already. If the lengths are different then
the test will fail and the if statement will never be reached so, remove
the if statements as they will only be reached if the lengths are equal.