Fix kmeans algorithm.
If no observations are mapped to a cluster then that centroid(codebook
entry) needs to be removed before moving on to the next update step.
This was not happening prior to this commit because the centroid was not
getting set to empty if no observations were found so the empty centroid
culling call was never finding any empty centoids meaning we always
maintained the full codebook(centroid list).