Update to consistent naming convention.
Everything is now talked about in terms of points where points have d
dimensions. Before this there was inconsistency between using points an
observations. Also, rename vq to a more reasonable function name. All it
does is identify clusters and comput distances. It is not, in of itself,
a full vector quantization algorithm. It's really just the encoder with
kmeans being the decoder and kmeans_optm putting it all together in the
return result.