Implement QueryProcessor class
This is the first (rudimentary) component for the query processing
pipeline. It is intended to handle the boilerplate passing objects,
arguments and data between Avocado's APIs.

This abstraction enables consumers of Avocado's APIs to rely on a
defined query processing _pipeline_ for parsing, interpreting,
augmenting, constructing, executing and presenting/packaging queries.

This introduces a new settings `QUERY_PROCESSORS` which is a dict of
alias => path items. `path` is a fully qualified import path to the
QueryProcessor class. `alias` is shorthand for referencing the class
in the `avocado.query.pipeline.query_processors` dict-like registry.