Add more tests to increase code coverage for kmeans package
The kmeans package is covered except for a single branch in the
`weighted_counts` method. I've had some difficulty reaching that branch
because the distances returned from `kmeans_optm` are never quite 0
because of the floating point precision involved in that method
resulting in near 0 distances in all my test cases.