Add RevisionManager `get_for_model` and `latest_for_model` methods
This provides a means of getting revisions across all instances for a
model. This is now used as the based for all other manager looks for
objects and a models.

In addition, a QuerySet can be passed in place of a model to limit the
which objects are included in the set.