Allow empty view and context attrs in query parser
These should not be checked at the query level. Instead take advantage
of all the logic that is already built into the view and context parsers
so this changes allows the query parser to take advantage of that by
always passing the context and view json to the respective parser
instead of ignoring it at the query level if it was empty. This commit
also reorders the queryset application to happen context first then
view. This change required some updates to the tests which are also
included in this version.