Remove parenthesis after now in accessed default value
With the parenthesis, the accessed value will always be the same because of the nature of Django defaults. With the parenthesis, accessed was always equal to datetime.now() whereas it is now set to datetime.now() when first saved(if no accessed value was passed) and then stays at that value until accessed is manually updated at a later time.