Change DATA_CACHE_ENABLED to default to False
There is [currently a bug](https://github.com/cbmi/avocado/issues/136) with `cached_methods`. Rather than having this be `True` and requiring devs to set it to `False` to avoid the bug, we can just make it `False` by default to avoid the bug altogether until it can be fixed.