Replace built-in ObjectSet and SetObject classes with django-objectset
This is defined as an optional dependency. For backwards compatibility, the
avocado.sets.models module remains for now to not break existing imports.

One hook that can be defined is defining the `label_field` attribute on the
ObjectSet class to denote which field should be treated as the label of the
sets.

Fix #75