Add `DataField.init` class method
This is a convenience method for creating and validating
a new `DataField` instance. It is preferred over doing
`DataField(...)` since it validates the field and exists prior
to saving.

A dot-delimited string can be used such as `DataField.init('app.model.field')` in addition to specifying
each argument separately.

This returns an unsaved instance for further modification.