Add support for using pre-defined model field choices
If pre-defined choices are specified, they take precedence over the
supplementary that are defined and used in the `values`, `labels`,
and `codes` (and related) methods. This is primarily to ensure
consistency of ordering and pairing of data.

Any method that requires a database lookup such as the `*_list` methods
and `search` will remain the same.

Fix #157