Refactor formatter for model-derived typing
This improves formatter processing to consider the `DataConcept` and
`DataField` `type` fields introduced in 878a3139 and cddcd9b,
respectively. In addition, the `DataField`'s simple type is included as
a fallback prior to using the 'raw' format.

This removes the need for `Formatter.default_formats` and reduces the
number of the `preferred_formats` across the exporter classes.

Fix #242

Signed-off-by: Byron Ruth <b@devel.io>