Remove query.AbstractData* classes
These were originally defined as abstract classes to support alternate
context, view, and query implementations, however this has proven to not
be needed. The implementation has be move into the concrete classes
defined in avocado.models.

The `engines` package has been added which will hold the
backend-specific processing utilities.

In addition, the `sql` method on the classes have been dropped in favor
of calling the function explicitly in the admin. Although this is
Django-specific and pending removal, this was done for backwards
compatibility.

Signed-off-by: Byron Ruth <b@devel.io>