Add DataField.engine field
This is the first step in migrating away from a Django-only data model.
DataFields will eventually be declarations of fields in data models
across various engines. For this reason, the engine field has been added
to specify the engine the field will be processed by during query
execution.

This field may not be needed in the future since it could be inferred
based on data model the field is bound to. The *only* data model
currently is the Django models denoted by app_name, model_name, and
field_name natural key.

Signed-off-by: Byron Ruth <b@devel.io>