Update search tests to evaluate the results before assessing them
There is a bug in Whoosh that results in an incorrect count of results
unless the results have been materialized. This conclusion was *deduced*
by myself after running the tests against the Solr and Elasticsearch
Haystack backends, both of which produced the correct count of results.

There was an open issue describing this behavior, however it has been
closed as invalid since then because no one provided a reproducible
example (https://bitbucket.org/mchaput/whoosh/issue/97/search-index-contains-a-lot-of-duplicates).

Signed-off-by: Byron Ruth <b@devel.io>