Add utilities to support canceling queries
Named connections are created and added to the `django.db.connections`
handler dynamically. The Django cache is used to store the derived
database alias and the process ID for the connection. This ensures
other threads or processes can lookup the process ID by name and cancel
the query using an separate connection. The only exception being
SQLite which can be interrupted across connections since it is
serverless.

At this time, the Oracle backend is not yet supported.

Close #280

Signed-off-by: Byron Ruth <b@devel.io>