Implement Query class to wrap querysets in the QueryProcessor
The `Query` class wraps a queryset and takes an optional name for making
the query cancelable. This is integrated in the `QueryProcessor.get_iterable`
method which now returns a `Query` instance and takes an optional `name`
argument for the query. `Query` implements the iterator protocol, so it
will be backwards compatible with it's previous usage.

Signed-off-by: Byron Ruth <b@devel.io>