Return empty iterable from QueryProcesor.get_iterable() for Django 1.5
An EmptyQuerySet in Django 1.5 and below was not acknowledged by the
internal query object, if used directly. This check ensures no results
are returned if working with an empty queryset.

In addition, the use of ModelTreeQuerySet has been removed in the
DataView parser to prevent overwriting the class if the input is an
EmptyQuerySet already.

Fix #279

Signed-off-by: Byron Ruth <b@devel.io>