Support configuring which databases are enabled during the test suite
Environment variables are the primary interface. The variables ENABLE_SQLITE,
ENABLE_MYSQL, ENABLE_POSTGRES can be set to either 1 or 0 to enable or
disable the database from the test suite. By default, sqlite is enabled
and the rest are disabled. In addition, the DEFAULT_ENGINE variable can
be used to set which database should be the primary for all tests (not
just database-specific tests). If only one database is used, it will be
selected as the default database, other sqlite is assumed.

Command line arguments can be passed to `test_suite.py` for convenience.
Use the format `--ENGINE` to enable a database and `--no-ENGINE` to
disable a database. `--default-engine=ENGINE` can be used to set the
default engine.

Signed-off-by: Byron Ruth <b@devel.io>