Add header check to base exporter class
The check uses the first row of data as input to validate the
shape of the header and row are the same.

The motivation behind this is to provide a bit more transparency
when an error is encountered during an export due to faulty
formatter.

In addition, the `to_FOO` formatter methods have been changed to
not take arbitrary positional or keyword arguments. The formatter
context represents the keyword arguments passed into the exporter
read method; typically included `request` and/or `user`.

Signed-off-by: Byron Ruth <b@devel.io>