Ignore caching for anonymous instances
The current implementation of the cache key depends on the instance
having a primary key. However in some cases (like the session), the
instance is temporary and not persisted. Adding this condition simply
ignores attempting to cache the data with respect to the key since
it is temporary and thus likely to change be short-lived in practice.

Signed-off-by: Byron Ruth <b@devel.io>